+++ src/main/java/dev/pcvolkmer/mv64e/datamapper/mapper/exceptionhandler/tuples/Tuple4.java
@@ -1,0 +1,34 @@
+/*
+ * This file is part of mv64e-onkostar-data
+ *
+ * Copyright (C) 2025  Paul-Christian Volkmer
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package dev.pcvolkmer.mv64e.datamapper.mapper.exceptionhandler.tuples;
+
+public class Tuple4<A, B, C, D> extends Tuple3<A, B, C> {
+  private final D value4;
+
+  public Tuple4(A value1, B value2, C value3, D value4) {
+    super(value1, value2, value3);
+    this.value4 = value4;
+  }
+
+  public D get4() {
+    return value4;
+  }
+}