+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.3.0](https://github.com/pcvolkmer/mv64e-onkostar-data/compare/v0.2.5...v0.3.0) (2025-12-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* update dto lib to version 0.2.0 ([#83](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/83))
+
+### deps
+
+* update dto lib to version 0.2.0 ([#83](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/83)) ([9a63d67](https://github.com/pcvolkmer/mv64e-onkostar-data/commit/9a63d6798a01f9afe2e350e0031040d207f02389))
+
+## [0.2.5](https://github.com/pcvolkmer/mv64e-onkostar-data/compare/v0.2.4...v0.2.5) (2025-12-15)
+
+
+### Features
+
+* add try-catch chain and related methods ([#77](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/77)) ([a28989c](https://github.com/pcvolkmer/mv64e-onkostar-data/commit/a28989cd73a31ea5d76526c7f9ab957a7945b7e7))
+* add tuple result values for try-catch chain ([#79](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/79)) ([8e4681d](https://github.com/pcvolkmer/mv64e-onkostar-data/commit/8e4681d54a7bc18a2454dd8a568bcf4174121b44))
+* ignore diagnosis mapping errors from [#63](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/63) ([#80](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/80)) ([e0126bf](https://github.com/pcvolkmer/mv64e-onkostar-data/commit/e0126bf5344ac327adef7eab6c9622c271806b66))
+* try/catch for procedures and therapies ([#82](https://github.com/pcvolkmer/mv64e-onkostar-data/issues/82)) ([00e0b47](https://github.com/pcvolkmer/mv64e-onkostar-data/commit/00e0b479ae4514215b8c800590d7d5b243fd227c))
+
 ## [0.2.4](https://github.com/pcvolkmer/mv64e-onkostar-data/compare/v0.2.3...v0.2.4) (2025-12-10)